+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/SolverLocalDB.java
@@ -3,6 +3,7 @@ package com.mageddo.dnsproxyserver.server.dns.solver;
 import com.mageddo.dnsproxyserver.config.Config;
 import com.mageddo.dnsproxyserver.config.Config.Entry.Type;
 import com.mageddo.dnsproxyserver.config.ConfigDAO;
+import com.mageddo.dnsproxyserver.config.Types;
 import com.mageddo.dnsproxyserver.server.dns.Messages;
 import dagger.Lazy;
 import lombok.AllArgsConstructor;
@@ -28,8 +29,8 @@ public class SolverLocalDB implements Solver {
     final var stopWatch = StopWatch.createStarted();
 
     final var type = Messages.findQuestionTypeCode(query);
-    if (Type.isNot(type, Type.A, Type.CNAME, Type.AAAA)) {
-      log.trace("status=typeNotSupported, action=continue, type={}, time={}", type, stopWatch.getTime());
+    if (Types.isNot(type, Type.A, Type.CNAME, Type.AAAA)) {
+      log.debug("status=typeNotSupported, action=continue, type={}, time={}", type, stopWatch.getTime());
       return null;
     }